+++ test/unit/regressions_test.rb
@@ -1,0 +1,11 @@
+require 'test_helper'
+
+class RegressionsTest < ActiveSupport::TestCase
+
+  test "XLSX" do
+  end
+
+  test "ODS" do
+  end
+
+end